requirements.txt
pyserial~=3.5
Flask~=1.1.2